refactor: remove undersores from some modules
